WIP - Add a vagrant file
This is the start of a vagrant file contribution.  In an ideal portable
development environment, a developer should only need to run the
following command.

    vagrant up

One should provision a fully running example development environment
bootstrapped.

Some documentation was removed from the Vagrantfile.  I created the
first Vagrantfile with the following command.

    vagrant init ubuntu/xenial64

- Learn more at https://www.vagrantup.com/
- Relates to #196